Split Webui::Packages::BuildLogController
Isolate the build log functionality into it's own controller.
