+++ src/api/.rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude --exclude-limit 100`
-# on 2023-09-21 12:17:09 UTC using RuboCop version 1.56.3.
+# on 2023-09-22 09:46:34 UTC using RuboCop version 1.56.3.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -42,6 +42,14 @@ FactoryBot/FactoryAssociationWithStrategy:
 # Offense count: 1
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
+# SupportedStyles: around, only_before
+Layout/EmptyLinesAroundAccessModifier:
+  Exclude:
+    - 'app/controllers/webui/packages/build_log_controller.rb'
+
+# Offense count: 1
+# This cop supports safe autocorrection (--autocorrect).
+# Configuration parameters: EnforcedStyle.
 # SupportedStyles: empty_lines, empty_lines_except_namespace, empty_lines_special, no_empty_lines, beginning_only, ending_only
 Layout/EmptyLinesAroundClassBody:
   Exclude:
@@ -58,6 +66,14 @@ Layout/LineContinuationLeadingSpace:
     - 'db/migrate/20180516074538_explicitly_set_charset.rb'
     - 'spec/controllers/webui/project_controller_spec.rb'
     - 'test/functional/source_controller_test.rb'
+
+# Offense count: 1
+# This cop supports safe autocorrection (--autocorrect).
+# Configuration parameters: EnforcedStyle, IndentationWidth.
+# SupportedStyles: aligned, indented
+Layout/LineEndStringConcatenationIndentation:
+  Exclude:
+    - 'app/controllers/webui/packages/build_log_controller.rb'
 
 # Offense count: 7
 # Configuration parameters: AllowedMethods.
@@ -173,7 +189,7 @@ Metrics/BlockNesting:
     - 'lib/xpath_engine.rb'
     - 'test/node_matcher.rb'
 
-# Offense count: 83
+# Offense count: 84
 # Configuration parameters: CountComments, Max, CountAsOne.
 Metrics/ClassLength:
   Exclude:
@@ -187,6 +203,7 @@ Metrics/ClassLength:
     - 'app/controllers/webui/comments_controller.rb'
     - 'app/controllers/webui/kiwi/images_controller.rb'
     - 'app/controllers/webui/package_controller.rb'
+    - 'app/controllers/webui/packages/build_log_controller.rb'
     - 'app/controllers/webui/patchinfo_controller.rb'
     - 'app/controllers/webui/project_controller.rb'
     - 'app/controllers/webui/projects/status_controller.rb'
@@ -282,6 +299,7 @@ Metrics/CyclomaticComplexity:
     - 'app/controllers/webui/monitor_controller.rb'
     - 'app/controllers/webui/package_controller.rb'
     - 'app/controllers/webui/packages/branches_controller.rb'
+    - 'app/controllers/webui/packages/build_log_controller.rb'
     - 'app/controllers/webui/packages/files_controller.rb'
     - 'app/controllers/webui/patchinfo_controller.rb'
     - 'app/controllers/webui/project_controller.rb'
@@ -1017,7 +1035,7 @@ Rails/FindEach:
     - 'db/data/20210707102506_add_api_endpoints_to_payload.rb'
     - 'db/data/20230620110143_backfill_scm_vendor_and_hook_event_in_workflow_run.rb'
 
-# Offense count: 27
+# Offense count: 28
 # Configuration parameters: Include.
 # Include: app/models/**/*.rb
 Rails/HasManyOrHasOneDependent:
@@ -1041,7 +1059,7 @@ Rails/HelperInstanceVariable:
     - 'app/helpers/webui/project_helper.rb'
     - 'app/helpers/webui/webui_helper.rb'
 
-# Offense count: 37
+# Offense count: 38
 # Configuration parameters: IgnoreScopes, Include.
 # Include: app/models/**/*.rb
 Rails/InverseOf:
@@ -1229,7 +1247,7 @@ Style/ConditionalAssignment:
   Exclude:
     - 'app/models/bs_request_action_submit.rb'
 
-# Offense count: 1000
+# Offense count: 1004
 # Configuration parameters: AllowedConstants.
 Style/Documentation:
   Enabled: false
@@ -1241,7 +1259,7 @@ Style/ExplicitBlockArgument:
     - 'app/lib/backend/connection.rb'
     - 'app/models/project.rb'
 
-# Offense count: 1741
+# Offense count: 1744
 # This cop supports unsafe autocorrection (--autocorrect-all).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: always, always_true, never
@@ -1365,14 +1383,7 @@ Style/SoleNestedConditional:
     - 'script/reformat_memprof'
     - 'test/node_matcher.rb'
 
-# Offense count: 1
-# This cop supports unsafe autocorrection (--autocorrect-all).
-# Configuration parameters: Mode.
-Style/StringConcatenation:
-  Exclude:
-    - 'app/controllers/concerns/webui/project_build_result_parsing.rb'
-
-# Offense count: 377
+# Offense count: 376
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: MinSize.
 # SupportedStyles: percent, brackets
