+++ src/api/app/views/event_mailer/build_fail.text.erb
@@ -1,4 +1,4 @@
-Visit <%= url_for(controller: 'webui/package', action: :live_build_log, project: event['project'], package: event['package'], repository: event['repository'], arch: event['arch'], only_path: false, host: @host) %>
+Visit <%= url_for(controller: 'webui/packages/build_log', action: :live_build_log, project: event['project'], package: event['package'], repository: event['repository'], arch: event['arch'], only_path: false, host: @host) %>
 
 Package <%= event['project'] %>/<%= event['package']%> failed to build in <%= event['repository']%>/<%= event['arch']%>
