+++ src/api/app/views/webui/package/_breadcrumb_items.html.haml
@@ -31,9 +31,6 @@
   - if current_page?(package_show_path(@project, @package))
     %li.breadcrumb-item.active{ 'aria-current' => 'page' }
       Overview
-  - if controller_name == 'package' && action_name == 'live_build_log'
-    %li.breadcrumb-item.active{ 'aria-current' => 'page' }
-      Build Log
   - if controller_name == 'package' && action_name == 'statistics'
     %li.breadcrumb-item
       = link_to 'Repository State', project_package_repository_binaries_path(@project, @package_name, @repository)
