+++ src/api/app/views/webui/packages/build_log/_breadcrumb_items.html.haml
@@ -1,0 +1,5 @@
+%li.breadcrumb-item.text-word-break-all
+  %i.fa.fa-archive
+  = link_to(@package, package_show_path(@project, @package))
+  %li.breadcrumb-item.active{ 'aria-current' => 'page' }
+    Build Log
