+++ src/api/app/views/webui/packages/build_log/_live_build_log_controls.html.haml
