+++ src/api/app/views/webui/packages/build_log/live_build_log.html.haml
@@ -2,7 +2,7 @@
 - @metarobots = 'noindex,nofollow'
 
 .card.mb-3
-  = render partial: 'tabs', locals: { project: @project, package: @package }
+  = render partial: 'webui/package/tabs', locals: { project: @project, package: @package }
   .card-body
     %h3= @pagetitle
     %p
