+++ src/api/app/views/webui/packages/build_log/update_build_log.js.haml
