+++ src/api/config/routes/webui_routes.rb
@@ -59,6 +59,13 @@ OBSApi::Application.routes.draw do
 
     resources :package, only: [:index], controller: 'webui/package', constraints: cons
 
+    controller 'webui/packages/build_log' do
+      get 'package/live_build_log/:project/:package/:repository/:arch' => :live_build_log, constraints: cons, as: 'package_live_build_log'
+      defaults format: 'js' do
+        get 'package/update_build_log/:project/:package/:repository/:arch' => :update_build_log, constraints: cons, as: 'package_update_build_log'
+      end
+    end
+
     controller 'webui/package' do
       defaults format: 'js' do
         get 'package/edit/:project/:package' => :edit, constraints: cons, as: 'edit_package'
@@ -90,9 +97,7 @@ OBSApi::Application.routes.draw do
         post 'package/save_group/:project/:package' => :save_group, constraints: cons, as: 'package_save_group'
         post 'package/remove_role/:project/:package' => :remove_role, constraints: cons, as: 'package_remove_role'
         get 'package/view_file/:project/:package/(:filename)' => :view_file, constraints: cons, as: 'package_view_file'
-        get 'package/live_build_log/:project/:package/:repository/:arch' => :live_build_log, constraints: cons, as: 'package_live_build_log'
         defaults format: 'js' do
-          get 'package/update_build_log/:project/:package/:repository/:arch' => :update_build_log, constraints: cons, as: 'package_update_build_log'
           post 'package/trigger_rebuild/:project/:package' => :trigger_rebuild, constraints: cons, as: 'package_trigger_rebuild'
           get 'package/abort_build/:project/:package' => :abort_build, constraints: cons, as: 'package_abort_build'
           post 'package/trigger_services/:project/:package' => :trigger_services, constraints: cons, as: 'package_trigger_services'
